backend/requirements.txt
google-generativeai
python-docx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask
flask-cors
werkzeug
requests
mangum

